Updted README.md and demo
